requirements.txt
Flask==3.1.0
flask-compress
gunicorn
psycopg2-binary
requests
python-dotenv
